+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Improvement: Include all default blueprints to `api/blueprints` folder;
 - Improvement: Rewrite default blueprints for CRUD and simplify them;
 - Improvement: Returns meta information for `GET /v1/:model` in body and `Content-Range` in headers;
+- Improvement: Add questions for Facebook and Twitter tokens;
 
 ## Version 0.3.0